+++ src/com/bretth/osmosis/core/mysql/v0_5/impl/ChangeWriter.java
@@ -233,7 +233,7 @@ public class ChangeWriter {
 			insertNodeStatement.setTimestamp(prmIndex++, new Timestamp(node.getTimestamp().getTime()));
 			insertNodeStatement.setInt(prmIndex++, fixedPrecisionConvertor.convertToFixed(node.getLatitude()));
 			insertNodeStatement.setInt(prmIndex++, fixedPrecisionConvertor.convertToFixed(node.getLongitude()));
-			insertNodeStatement.setInt(prmIndex++, tileCalculator.calculateTile(node.getLatitude(), node.getLongitude()));
+			insertNodeStatement.setLong(prmIndex++, tileCalculator.calculateTile(node.getLatitude(), node.getLongitude()));
 			insertNodeStatement.setString(prmIndex++, tagFormatter.format(node.getTagList()));
 			insertNodeStatement.setBoolean(prmIndex++, visible);
 			insertNodeStatement.setLong(prmIndex++, userIdManager.getUserId());
@@ -261,7 +261,7 @@ public class ChangeWriter {
 			insertNodeCurrentStatement.setTimestamp(prmIndex++, new Timestamp(node.getTimestamp().getTime()));
 			insertNodeCurrentStatement.setInt(prmIndex++, fixedPrecisionConvertor.convertToFixed(node.getLatitude()));
 			insertNodeCurrentStatement.setInt(prmIndex++, fixedPrecisionConvertor.convertToFixed(node.getLongitude()));
-			insertNodeCurrentStatement.setInt(prmIndex++, tileCalculator.calculateTile(node.getLatitude(), node.getLongitude()));
+			insertNodeCurrentStatement.setLong(prmIndex++, tileCalculator.calculateTile(node.getLatitude(), node.getLongitude()));
 			insertNodeCurrentStatement.setString(prmIndex++, tagFormatter.format(node.getTagList()));
 			insertNodeCurrentStatement.setBoolean(prmIndex++, visible);
 			insertNodeCurrentStatement.setLong(prmIndex++, userIdManager.getUserId());
